Some cleanup and improvements
