+++ packages/react-native-reanimated/src/common/processors/index.ts
@@ -1,5 +1,4 @@
 'use strict';
 export * from './colors';
 export * from './shadows';
-export * from './shadows.web';
 export * from './transformOrigin';
