--- packages/react-native-reanimated/src/common/processors/shadows.web.ts
@@ -1,46 +1,0 @@
-'use strict';
-import type { BoxShadowValue, ViewStyle } from 'react-native';
-
-import type { ValueProcessor } from '../types';
-import { maybeAddSuffix, parseBoxShadowString } from '../utils';
-
-export const processBoxShadowWeb: ValueProcessor<
-  string | ReadonlyArray<BoxShadowValue>,
-  Record<string, string> | string
-> = (value) => {
-  const parsedShadow =
-    typeof value === 'string' ? parseBoxShadowString(value) : value;
-
-  return parsedShadow
-    .map(
-      ({
-        offsetX,
-        offsetY,
-        color = '#000',
-        blurRadius = '',
-        spreadDistance = '',
-        inset = '',
-      }) =>
-        [
-          maybeAddSuffix(offsetX, 'px'),
-          maybeAddSuffix(offsetY, 'px'),
-          maybeAddSuffix(blurRadius, 'px'),
-          maybeAddSuffix(spreadDistance, 'px'),
-          color,
-          inset ? 'inset' : '',
-        ]
-          .filter(Boolean)
-          .join(' ')
-    )
-    .join(', ');
-};
-
-type ShadowOffset = NonNullable<ViewStyle['shadowOffset']>;
-
-export const processShadowOffset: ValueProcessor<
-  ShadowOffset,
-  Record<string, string> | string
-> = (value) => ({
-  width: `${value.width}px`,
-  height: `${value.height}px`,
-});
