+++ packages/react-native-reanimated/src/common/utils/suffix.ts
@@ -1,6 +1,6 @@
 'use strict';
 
-function hasSuffix(value: unknown): value is string {
+export function hasSuffix(value: unknown): value is string {
   return typeof value === 'string' && isNaN(parseInt(value[value.length - 1]));
 }
