+++ packages/react-native-reanimated/src/css/platform/web/style/builders/shadows.ts
@@ -1,10 +1,21 @@
 'use strict';
 import type { TextStyle, ViewStyle } from 'react-native';
 
-import { logger, processShadowOffset } from '../../../../../common';
+import type { ValueProcessor } from '../../../../../common';
+import { logger } from '../../../../../common';
 import { opacifyColor } from '../../utils';
 import { createRuleBuilder } from '../builderFactories';
 import { processColor } from '../processors';
+
+type ShadowOffset = NonNullable<ViewStyle['shadowOffset']>;
+
+const processShadowOffset: ValueProcessor<
+  ShadowOffset,
+  Record<string, string> | string
+> = (value) => ({
+  width: `${value.width}px`,
+  height: `${value.height}px`,
+});
 
 type BoxShadowProps = Pick<
   ViewStyle,
